+++ src/components/NotFound.js
@@ -1,9 +1,9 @@
 import React from 'react'
 
 const NotFound = () => (
-  <div className='p2 center'>
-    <h1 className='mb1 monospace'>404</h1>
-    <p>Sorry, we can’t seem to find that page.</p>
+  <div className='mb8 px2 py8 center'>
+    <h1 className='mt0 mb1 monospace'>404</h1>
+    <p className='m0 h4'>Sorry, we can’t seem to find that page.</p>
   </div>
 )
